Fix: Convert time to unsigned long long

+++ Version Base/Functions/Physics/CellWM/AgentsSystem.h
@@ -28,7 +28,7 @@ public:
 
 	float       _radius;
 
-	double      _itTime;
+	unsigned long long  _itTime;
 
 	CellWM		*_agentsWM_DFC;
 public:
